Adding all the logging.
This currently fails because on the second attempt to backup, the backup function finds that this backup already exists in the object storage.

If we want the retry to work we either need to clean-up the first attempt or decide that it's ok that it already exists.

To say we're doing a retry, we could add an argument to the backup function and change the if statement appropriately.

However, I don't understand the larger context well enough to know what impact that would have.